refactor(transcription): consolidate constants into dedicated file
